Added "ember-decouple" module and updated codebase to use this module
Ember-decouple is a very basic plugin that checks to see if the target
platform is windows 8. If it is windows 8, then innerHTML and $.append
methods are customized to work around the windows html1701 error.

Replaced `innerHTML` with `Ember.decouple.innerHTML(target, content)`.
By proxying these "dangerous" calls to ember-decouple, ember.js is
now usable inside a windows8 application.